+++ services/cdn/mgmt/2021-06-01/cdn/afdendpoints.go
@@ -430,7 +430,8 @@ func (client AFDEndpointsClient) ListByProfileComplete(ctx context.Context, reso
 	return
 }
 
-// ListResourceUsage checks the quota and actual usage of endpoints under the given CDN profile.
+// ListResourceUsage checks the quota and actual usage of the given AzureFrontDoor endpoint under the given CDN
+// profile.
 // Parameters:
 // resourceGroupName - name of the Resource group within the Azure subscription.
 // profileName - name of the Azure Front Door Standard or Azure Front Door Premium profile which is unique
@@ -751,7 +752,8 @@ func (client AFDEndpointsClient) UpdateResponder(resp *http.Response) (result AF
 	return
 }
 
-// ValidateCustomDomain validates the custom domain mapping to ensure it maps to the correct CDN endpoint in DNS.
+// ValidateCustomDomain validates the custom domain mapping to ensure it maps to the correct CDN endpoint in DNS.This
+// api isn't work for apex domain.
 // Parameters:
 // resourceGroupName - name of the Resource group within the Azure subscription.
 // profileName - name of the Azure Front Door Standard or Azure Front Door Premium profile which is unique