+++ services/cdn/mgmt/2021-06-01/cdn/afdorigingroups.go
@@ -440,7 +440,8 @@ func (client AFDOriginGroupsClient) ListByProfileComplete(ctx context.Context, r
 	return
 }
 
-// ListResourceUsage checks the quota and actual usage of endpoints under the given CDN profile.
+// ListResourceUsage checks the quota and actual usage of the given AzureFrontDoor origin group under the given CDN
+// profile.
 // Parameters:
 // resourceGroupName - name of the Resource group within the Azure subscription.
 // profileName - name of the Azure Front Door Standard or Azure Front Door Premium profile which is unique